# specs/001-portfolio-performance-dashboard/contracts/api.yaml
openapi: 3.0.3
info:
  title: Portfolio Management API
  description: RESTful API for intelligent portfolio management system
  version: 0.1.0
  contact:
    name: Portfolio Management System
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - ApiKeyAuth: []
paths:
...
  /portfolios:
    get:
      tags: [Portfolios]
      summary: List user portfolios
      operationId: listPortfolios
      responses:
        '200':
          description: List of portfolios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Portfolio'
    
    post:
      tags: [Portfolios]
      summary: Create new portfolio
      operationId: createPortfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
              required: [name]
      responses:
        '201':
          description: Portfolio created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Portfolio'
  /portfolios/{portfolio_id}:
    get:
      tags: [Portfolios]
      summary: Get portfolio details
      operationId: getPortfolio
      parameters:
        - name: portfolio_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Portfolio details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioDetail'
        '404':
          description: Portfolio not found

    put:
      tags: [Portfolios]
      summary: Update portfolio
      operationId: updatePortfolio
      parameters:
        - name: portfolio_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Portfolio updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Portfolio'

    delete:
      tags: [Portfolios]
      summary: Delete portfolio
      operationId: deletePortfolio
      parameters:
        - name: portfolio_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Portfolio deleted

  # Holdings
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Portfolio:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        total_value:
          type: number
          format: decimal
        daily_change:
          type: number
          format: decimal
        daily_change_percent:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PortfolioDetail:
      allOf:
        - $ref: '#/components/schemas/Portfolio'
        - type: object
          properties:
            holdings:
              type: array
              items:
                $ref: '#/components/schemas/Holding'
            recent_transactions:
              type: array
              items:
                $ref: '#/components/schemas/Transaction'
    Holding:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stock:
          $ref: '#/components/schemas/Stock'
        quantity:
          type: number
          format: decimal
        average_cost:
          type: number
          format: decimal
        current_value:
          type: number
          format: decimal
        unrealized_gain_loss:
          type: number
          format: decimal
        unrealized_gain_loss_percent:
          type: number
          format: decimal
        updated_at:
          type: string
          format: date-time
    Stock:
      type: object
      properties:
        id:
          type: string
          format: uuid
        symbol:
          type: string
        company_name:
          type: string
        exchange:
          type: string
        current_price:
          type: number
          format: decimal
        daily_change:
          type: number
          format: decimal
        daily_change_percent:
          type: number
          format: decimal
        status:
          type: string
          enum: [ACTIVE, HALTED, SUSPENDED, DELISTED]
        last_price_update:
          type: string
          format: date-time
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stock:
          $ref: '#/components/schemas/Stock'
        transaction_type:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: number
          format: decimal
        price_per_share:
          type: number
          format: decimal
        total_amount:
          type: number
          format: decimal
        fees:
          type: number
          format: decimal
        transaction_date:
          type: string
          format: date
        source_type:
          type: string
          enum: [EMAIL, MANUAL, API]
        notes:
          type: string
        is_verified:
          type: boolean
        processed_date:
          type: string
          format: date-time
...